requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
python-decouple
Pillow
gunicorn
whitenoise
dj-database-url
django-allauth
psycopg2-binary

# Neon PostgreSQL
psycopg[binary]
neon-api

# Cloudflare R2 (S3-compatible)
django-storages[s3]
boto3

# PDF/Export generation
reportlab
qrcode[pil]
xhtml2pdf
